+++ lib/src/gpio_const.dart
@@ -1,26 +1,25 @@
-
 /// Mapping from physical pin on RPi model 2 and 3 to BCM GPIO.
 const physToBcmGpioRPi2 = const <int>[
-  -1,     // 0
+  -1, //     0
   -1, -1, // 1, 2
-   2, -1,
-   3, -1,
-   4, 14,
+  02, -1,
+  03, -1,
+  04, 14,
   -1, 15,
   17, 18,
   27, -1,
   22, 23,
   -1, 24,
   10, -1,
-   9, 25,
-  11,  8,
-  -1,  7, // 25, 26
+  09, 25,
+  11, 08,
+  -1, 07, // 25, 26
 
 // B+
 
-   0,  1, // 27, 28
-   5, -1,
-   6, 12,
+  00, 01, // 27, 28
+  05, -1,
+  06, 12,
   13, -1,
   19, 16,
   26, 20,
@@ -41,44 +40,20 @@ const physToBcmGpioRPi2 = const <int>[
   -1, -1,
 ];
 
-/// Mapping from physical pin on RPi model 2 and 3 to PWM port.
-const physToPwmPort = const <int>[
-  -1, // 0
-  -1, -1, // 1, 2
-  -1, -1,
-  -1, -1,
-  -1, -1,
-  -1, -1,
-  -1,  0, // 11, 12
-   1, -1,
-  -1, -1,
-  -1,  0, // 17, 18
-   1, -1,
-  -1, -1,
-  -1, -1,
-  -1, -1, // 25, 26
+const physI2CPins = <int>[
+  3, // I2C SDA
+  5, // I2C SCL
+];
 
-// B+
+const physSpiPins = <int>[
+  19, // SPI0 MOSI
+  21, // SPI0 MISO
+  23, // SPI0 SCLK
+  24, // SPI0 CS0
+  26, // SPI0 CS1
+];
 
-  -1, -1, // 27, 28
-  -1, -1,
-  -1, -1,
-  -1, -1,
-  -1, -1,
-  -1, -1,
-  -1,  0, // 39, 40
-
-// the P5 connector on the Rev 2 boards:
-
-   1, -1,
-  -1, -1,
-   1, -1,
-  -1, -1,
-  -1, -1,
-  -1, -1,
-  -1, -1,
-  -1, -1,
-  -1, -1,
-  -1, -1,
-  -1, -1,
+const physEepromPins = <int>[
+  27, // Reserved
+  28, // Reserved
 ];